latest version

+++ app/Http/Controllers/UploadController.php
@@ -7,11 +7,11 @@ use Chumper\Zipper\Zipper;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Storage;
-
+use Imagick;
 class UploadController extends Controller
 {
 
-    protected $fieldUpload = ['id', 'title', 'description', 'agreement', 'cover_file', 'author', 'user_id', 'isPublic', 'fileUrl', 'group_id', 'cat_id', 'view', 'sub_cat', 'grade','link_test'];
+    protected $fieldUpload = ['id', 'title', 'description', 'agreement', 'cover_file', 'author', 'user_id', 'isPublic', 'fileUrl', 'group_id', 'cat_id', 'view', 'sub_cat', 'grade','link_pretest', 'link_test', 'type_book'];
     protected $fieldGroup = ['group_id', 'book_id'];
 
     /**
@@ -55,26 +55,51 @@ class UploadController extends Controller
     // @return Respose
     public function upload(Request $request)
     {
-        if ($_FILES["bookfile"]["error"] > 0)
-       {
-          var_dump($_FILES["bookfile"]["error"]);
-          return;
-       }
+        $lastBookID = $this->getNextBookInsertID();
 
-	$lastBookID = $this->getNextBookInsertID();
-	
+        if($request->type_book == 0) { // E-Book
         $file = $request->file('bookfile')->store($lastBookID, 'book');
         $fullpath = Storage::disk('book')->path('/');
-        $extract_dir = $fullpath . $lastBookID;
         chmod($fullpath . $lastBookID, 0777);
         //Extract files
+       
 
-        $zipper = new Zipper;
+           $extract_dir = $fullpath . $lastBookID;
+           $zipper = new Zipper;
+   
+           $zipper->make($fullpath . $file)->folder('')->extractTo($extract_dir);
+   
+           $contents = scandir($extract_dir);
+           // check if $contents is a directory and actually has items
+           if (is_array($contents) && count($contents)) {
+               foreach($contents as $item) { // loop through directory contents
+                   if (substr(strtolower($item), -5) == ".html") { // checking if a file ends with .html
+                   rename($extract_dir . "/" . $item, $extract_dir . "/index.html"); // rename it to index.html
+                   break; // no need to loop more, job's done
+                   }
+               }
+           }
+   
+           unlink($fullpath . $file);
+           $zipper->close();
 
-        $zipper->make($fullpath . $file)->folder('')->extractTo($extract_dir);
+           $mainfile = $lastBookID . '/index.html';
+           $coverfile = $lastBookID . '/files/large/1.jpg';
 
-        unlink($fullpath . $file);
-        $zipper->close();
+        } else { // PDF
+            $md5Name = md5_file($request->file('pdf_bookfile')->getRealPath());
+            $guessExtension = $request->file('pdf_bookfile')->guessExtension();
+
+            $pdf_store_file = $request->file('pdf_bookfile')->storeAs($lastBookID, $md5Name.'.'.$guessExtension  ,'book');
+            $pdf_path = Storage::disk('book')->path($pdf_store_file);
+
+            $im = new Imagick($pdf_path.'[0]');
+            $im->setImageFormat('jpg');
+            $coverfile = $lastBookID.'/cover.jpg';
+            $mainfile = $lastBookID . '/' . $md5Name.'.'.$guessExtension;
+            file_put_contents(Storage::disk('book')->path($coverfile), $im);
+        }
+
         if (!isset($request->isPublic)) {
             $request->isPublic = 0;
         }
@@ -82,8 +107,7 @@ class UploadController extends Controller
         $userid = ($request->userid == '') ? Auth::user()->id : $request->userid;
         $group = ($request->group == '') ? 0 : $request->group;
 
-        $mainfile = $lastBookID . '/index.html';
-        $coverfile = $lastBookID . '/files/large/1.jpg';
+
         $link_test = $request->attachment;
 
         $bookdata = array_combine($this->fieldUpload, [
@@ -94,13 +118,17 @@ class UploadController extends Controller
             $coverfile, 
             $request->author, 
             $userid, 
-            $request->isPublic, 
-            $mainfile, $group, 
+            $request->isGlobal,
+            $mainfile, 
+            $group, 
             $request->category, 
             0, 
             $request->sub, 
             $request->grade, 
-            $link_test]);
+            $request->before_attachment,
+            $link_test,
+            $request->type_book
+            ]);
         $groupdata = array_combine($this->fieldGroup, [$group, $lastBookID]);
         // print_r($bookdata);
         // exit;
